rename: wish DTO package이름 코드 일관성에 맞게 변경
